// src/models/patient.ts
export interface patient {
    key? : string;
    name : string;
    dob: Date;
    nric: string;
    medicalhistory: string;
    annualincome: Float32Array;
    userid?: string;
    address: string;
    relationship? : string;
    postalcode: number;
    phone: number;
    gender: string;
    patientid?: string;


}